requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
auth0-python==4.7.2
Authlib==1.3.2
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Flask==3.0.3
Flask-Caching==2.3.0
Flask-OAuthlib==0.9.6
fonttools==4.54.1
frozenlist==1.4.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
multidict==6.1.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
numpy==2.1.1
oauthlib==2.1.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pycparser==2.22
pydexcom==0.4.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-oauthlib==1.1.0
six==1.16.0
twilio==9.3.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.12.1
